Moved URI strategy outside of RDF mapping classes
