+++ lib/Connections2JSONLD.js
@@ -4,7 +4,7 @@
  */
 var Transform = require('stream').Transform,
     util = require('util'),
-    moment = require('moment');
+    URIStrategy = require('./URIStrategy.js');
 
 var Connections2JSONLD = function (baseUris, context) {
   Transform.call(this, {objectMode : true});
@@ -25,46 +25,24 @@ var Connections2JSONLD = function (baseUris, context) {
       "arrivalTime" : "http://semweb.mmlab.be/ns/linkedconnections#arrivalTime"
     } 
   };
-  
-  var defaultBaseUris = {
-    stops : 'http://example.org/stops/',
-    connections : 'http://example.org/connections/',
-    trips : 'http://example.org/trips/',
-    routes : 'http://example.org/routes/'
-  };
-  if (!baseUris) {
-    baseUris = defaultBaseUris;
-  } else {
-    if (typeof baseUris.stops !== 'string') {
-      baseUris.stops = defaultBaseUris.stops;
-    }
-    if (typeof baseUris.trips !== 'string') {
-      baseUris.trips = defaultBaseUris.trips;
-    }
-    if (typeof baseUris.routes !== 'string') {
-      baseUris.routes = defaultBaseUris.routes;
-    }
-    if (typeof baseUris.connections !== 'string') {
-      baseUris.connections = defaultBaseUris.connections;
-    }
-  }
-  this._baseUris = baseUris;
+
+  this._uris = new URIStrategy(baseUris);
   this._count = 0;
 };
 
 util.inherits(Connections2JSONLD, Transform);
 
 Connections2JSONLD.prototype._transform = function (connection, encoding, done) {
-  var id = this._baseUris.connections + encodeURIComponent(connection.departureTime + connection.departureStop + connection.trip);
+  var id = this._uris.getId(connection);
   done(null, {
     "@id" : id,
     "@type" : "Connection",
-    "departureStop" : this._baseUris.stops + encodeURIComponent(connection.departureStop),
-    "arrivalStop" : this._baseUris.stops + encodeURIComponent(connection.arrivalStop),
+    "departureStop" : this._uris.getStopId(connection.departureStop),
+    "arrivalStop" : this._uris.getStopId(connection.arrivalStop),
     "departureTime" : connection.departureTime.toISOString(),
     "arrivalTime" : connection.arrivalTime.toISOString(),
-    "gtfs:trip" : this._baseUris.trips + encodeURIComponent(connection.trip),
-    "gtfs:route" : this._baseUris.routes + encodeURIComponent(connection.route)
+    "gtfs:trip" : this._uris.getTripId(connection.trip),
+    "gtfs:route" : this._uris.getRouteId(connection.route)
   });
 };
