+++ lib/Connections2Triples.js
@@ -4,40 +4,18 @@
  */
 var Transform = require('stream').Transform,
     util = require('util'),
-    moment = require('moment');
+    URIStrategy = require('./URIStrategy.js');
 
 var Connections2Triples = function (baseUris) {
   Transform.call(this, {objectMode : true});
-  var defaultBaseUris = {
-    stops : 'http://example.org/stops/',
-    connections : 'http://example.org/connections/',
-    trips : 'http://example.org/trips/',
-    routes : 'http://example.org/routes/'
-  };
-  if (!baseUris) {
-    baseUris = defaultBaseUris;
-  } else {
-    if (typeof baseUris.stops !== 'string') {
-      baseUris.stops = defaultBaseUris.stops;
-    }
-    if (typeof baseUris.trips !== 'string') {
-      baseUris.trips = defaultBaseUris.trips;
-    }
-    if (typeof baseUris.routes !== 'string') {
-      baseUris.routes = defaultBaseUris.routes;
-    }
-    if (typeof baseUris.connections !== 'string') {
-      baseUris.connections = defaultBaseUris.connections;
-    }
-  }
-  this._baseUris = baseUris;
+  this._uris = new URIStrategy(baseUris);
   this._count = 0;
 };
 
 util.inherits(Connections2Triples, Transform);
 
 Connections2Triples.prototype._transform = function (connection, encoding, done) {
-  var id = this._baseUris.connections + encodeURIComponent(connection.departureTime + connection.departureStop + connection.trip);
+  var id = this._uris.getId(connection);
   this.push({
     subject : id,
     predicate :'http://www.w3.org/1999/02/22-rdf-syntax-ns#type',
@@ -46,12 +24,12 @@ Connections2Triples.prototype._transform = function (connection, encoding, done)
   this.push({
     subject : id,
     predicate :'http://semweb.mmlab.be/ns/linkedconnections#departureStop',
-    object : this._baseUris.stops + encodeURIComponent(connection.departureStop)
+    object : this._uris.getStopId(connection.departureStop)
   });
   this.push({
     subject : id,
     predicate :'http://semweb.mmlab.be/ns/linkedconnections#arrivalStop',
-    object : this._baseUris.stops + encodeURIComponent(connection.arrivalStop)
+    object : this._uris.getStopId(connection.arrivalStop)
   });
   this.push({
     subject : id,
@@ -66,12 +44,12 @@ Connections2Triples.prototype._transform = function (connection, encoding, done)
   this.push({
     subject : id,
     predicate :'http://vocab.gtfs.org/terms#trip',
-    object : this._baseUris.trips + encodeURIComponent(connection.trip)
+    object : this._uris.getTripId(connection)
   });
   this.push({
     subject : id,
     predicate :'http://vocab.gtfs.org/terms#route',
-    object : this._baseUris.routes + encodeURIComponent(connection.route)
+    object : this._uris.getRouteId(connection);
   });
   done();
 };
